ユーザーは半径 100 m 以内のレストランの開閉店時間一覧を見れる：Map コンポーネントを作成
